+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   YDF can load TF-DF models directly (i.e. a TF model with a YDF model in the
     "assets" sub directory).
+-   Expose confusion tables in a GBT model's analysis.
 
 ## 1.1.0 - 2022-10-21